apps/backend-api/FastAPI/requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
